+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 * Add associative table UserBooking - JC
 * Add more users - JC
 * Add Booking API - JC
+* Modify Get Booking API - JC
 
 #### Changed
 * Fix volunteer model - JC